Edited database.

+++ src/models/project.hpp
@@ -154,7 +154,7 @@ private:
     friend Mapper<Project>;
 #ifdef __cpp_impl_coroutine
     friend CoroMapper<Project>;
-#endif  // ifdef __cpp_impl_coroutine
+#endif // ifdef __cpp_impl_coroutine
     static const std::vector<std::string>& insertColumns() noexcept;
     void outputArgs(drogon::orm::internal::SqlBinder& binder) const;
     const std::vector<std::string> updateColumns() const;
@@ -208,7 +208,7 @@ public:
         if(parametersCount > 0) {
             sql[sql.length()-1]=')';
             sql += " values (";
-        } else {
+        } else   {
             sql += ") values (";
         }
 
@@ -217,11 +217,11 @@ public:
         size_t n=0;
         sql +="default,";
         if(dirtyFlag_[1]) {
-            n = static_cast<size_t>(sprintf(placeholderStr, "$%d,", placeholder++) );
+            n = sprintf(placeholderStr, "$%d,", placeholder++);
             sql.append(placeholderStr, n);
         }
         if(dirtyFlag_[2]) {
-            n = static_cast<size_t>(sprintf(placeholderStr, "$%d,", placeholder++) );
+            n = sprintf(placeholderStr, "$%d,", placeholder++);
             sql.append(placeholderStr, n);
         }
         if(parametersCount > 0)
